+++ src/main/raceTimeout.ts
@@ -1,4 +1,5 @@
-import { AbortableCallback } from './shared-types';
+import { AbortableCallback } from './types';
+import { isPromiseLike } from './utils';
 
 /**
  * Returns a promise that is rejected after the timeout elapses.
@@ -7,7 +8,14 @@ import { AbortableCallback } from './shared-types';
  * @param ms The timeout after which the promise is rejected with an {@link !Error Error}.
  * @returns The fulfilled value.
  */
-export function raceTimeout<T>(cb: AbortableCallback<T>, ms: number): Promise<T> {
+export function raceTimeout<T>(cb: AbortableCallback<T>, ms: number): Promise<T>;
+
+export function raceTimeout<T>(promise: PromiseLike<T>, ms: number): Promise<T>;
+
+export function raceTimeout<T>(cb: AbortableCallback<T> | PromiseLike<T>, ms: number): Promise<T> {
+  if (isPromiseLike(cb)) {
+    return raceTimeout(() => cb, ms);
+  }
   return new Promise((resolve, reject) => {
     const abortController = new AbortController();
